Make it obnoxiously hard to misuse the Token interface.
Token construction is only allowed through a `make` function template
that statically requires a token to be complex for it to hold a lexeme.
Lexeme access is also provided through a function template the provides
static and runtime checks that a lexeme is valid before supplying it
back to the caller.